refactor: following을 유저 도메인으로 합침
