+++ src/main/java/com/ddudu/user/domain/Following.java
@@ -1,9 +1,8 @@
-package com.ddudu.following.domain;
+package com.ddudu.user.domain;
 
 import com.ddudu.common.BaseEntity;
 import com.ddudu.common.exception.InvalidParameterException;
-import com.ddudu.following.exception.FollowingErrorCode;
-import com.ddudu.user.domain.User;
+import com.ddudu.user.exception.FollowingErrorCode;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
