+++ src/main/java/com/ddudu/user/service/FollowingService.java
@@ -1,25 +1,23 @@
-package com.ddudu.following.service;
+package com.ddudu.user.service;
 
 import com.ddudu.common.exception.DataNotFoundException;
 import com.ddudu.common.exception.DuplicateResourceException;
 import com.ddudu.common.exception.ForbiddenException;
 import com.ddudu.common.exception.InvalidParameterException;
-import com.ddudu.following.domain.Following;
-import com.ddudu.following.domain.Following.FollowingBuilder;
-import com.ddudu.following.domain.FollowingStatus;
-import com.ddudu.following.dto.request.FollowRequest;
-import com.ddudu.following.dto.request.UpdateFollowingRequest;
-import com.ddudu.following.dto.response.FollowingResponse;
-import com.ddudu.following.exception.FollowingErrorCode;
-import com.ddudu.following.repository.FollowingRepository;
+import com.ddudu.user.domain.Following;
+import com.ddudu.user.domain.Following.FollowingBuilder;
+import com.ddudu.user.domain.FollowingStatus;
 import com.ddudu.user.domain.Options;
 import com.ddudu.user.domain.User;
+import com.ddudu.user.dto.request.FollowRequest;
+import com.ddudu.user.dto.request.UpdateFollowingRequest;
+import com.ddudu.user.dto.response.FollowingResponse;
+import com.ddudu.user.exception.FollowingErrorCode;
+import com.ddudu.user.repository.FollowingRepository;
 import com.ddudu.user.repository.UserRepository;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.validation.annotation.Validated;
 
 @Service
 @RequiredArgsConstructor
@@ -29,16 +27,9 @@ public class FollowingService {
   private final UserRepository userRepository;
   private final FollowingRepository followingRepository;
 
-  private static void checkPermission(Long loginId, Following following) {
-    if (!following.isOwnedBy(loginId)) {
-      throw new ForbiddenException(FollowingErrorCode.WRONG_OWNER);
-    }
-  }
-
   @Transactional
-  @Validated
-  public FollowingResponse create(Long followerId, @Valid FollowRequest request) {
-    User follower = userRepository.findById(followerId)
+  public FollowingResponse create(Long loginId, FollowRequest request) {
+    User follower = userRepository.findById(loginId)
         .orElseThrow(() -> new DataNotFoundException(FollowingErrorCode.FOLLOWER_NOT_EXISTING));
     User followee = userRepository.findById(request.followeeId())
         .orElseThrow(() -> new DataNotFoundException(FollowingErrorCode.FOLLOWEE_NOT_EXISTING));
@@ -63,11 +54,13 @@ public class FollowingService {
   }
 
   @Transactional
-  public FollowingResponse updateStatus(Long id, Long followerId, UpdateFollowingRequest request) {
-    Following following = followingRepository.findById(id)
+  public FollowingResponse updateStatus(
+      Long loginId, Long followingId, UpdateFollowingRequest request
+  ) {
+    Following following = followingRepository.findById(followingId)
         .orElseThrow(() -> new DataNotFoundException(FollowingErrorCode.ID_NOT_EXISTING));
 
-    if (!following.isOwnedBy(followerId)) {
+    if (!following.isOwnedBy(loginId)) {
       throw new ForbiddenException(FollowingErrorCode.WRONG_OWNER);
     }
 
@@ -83,12 +76,18 @@ public class FollowingService {
   }
 
   @Transactional
-  public void delete(Long id, Long loginId) {
-    followingRepository.findById(id)
+  public void delete(Long loginId, Long followingId) {
+    followingRepository.findById(followingId)
         .ifPresent(following -> {
           checkPermission(loginId, following);
           followingRepository.delete(following);
         });
+  }
+
+  private void checkPermission(Long loginId, Following following) {
+    if (!following.isOwnedBy(loginId)) {
+      throw new ForbiddenException(FollowingErrorCode.WRONG_OWNER);
+    }
   }
 
 }
