+++ src/test/java/com/ddudu/user/service/FollowingServiceTest.java
@@ -1,4 +1,4 @@
-package com.ddudu.following.service;
+package com.ddudu.user.service;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
@@ -8,16 +8,17 @@ import com.ddudu.common.exception.BadRequestException;
 import com.ddudu.common.exception.DataNotFoundException;
 import com.ddudu.common.exception.DuplicateResourceException;
 import com.ddudu.common.exception.ForbiddenException;
-import com.ddudu.following.domain.Following;
-import com.ddudu.following.domain.FollowingStatus;
-import com.ddudu.following.dto.request.FollowRequest;
-import com.ddudu.following.dto.request.UpdateFollowingRequest;
-import com.ddudu.following.dto.response.FollowingResponse;
-import com.ddudu.following.exception.FollowingErrorCode;
-import com.ddudu.following.repository.FollowingRepository;
+import com.ddudu.user.domain.Following;
+import com.ddudu.user.domain.FollowingStatus;
+import com.ddudu.user.dto.request.FollowRequest;
+import com.ddudu.user.dto.request.UpdateFollowingRequest;
+import com.ddudu.user.dto.response.FollowingResponse;
+import com.ddudu.user.exception.FollowingErrorCode;
+import com.ddudu.user.repository.FollowingRepository;
 import com.ddudu.user.domain.Options;
 import com.ddudu.user.domain.User;
 import com.ddudu.user.repository.UserRepository;
+import com.ddudu.user.service.FollowingService;
 import java.util.Objects;
 import java.util.Optional;
 import net.datafaker.Faker;
